Finished big part of the calculations + gui for rocket and crafting

+++ src/main/scala/net/willemjan/factorio/model/FactorioData.scala
@@ -38,7 +38,7 @@ abstract class AbstractItem
 
 case class Item(name: String, subGroup: String, icon: String, placeResult: String, stackSize: Int, craftingCategories: Seq[String] = Seq.empty) extends AbstractItem
 case class Fluid(name: String) extends AbstractItem
-case class Recipe(name: String, category: Category, ingredients: Seq[Ingredient], requiredEnergy: Integer, results: Seq[RecipeResult]) extends AbstractItem {
+case class Recipe(name: String, category: Category, ingredients: Seq[Ingredient], requiredEnergy: Double, results: Seq[RecipeResult]) extends AbstractItem {
   def withIngredients(ingredients: Seq[Ingredient]): Recipe = this.copy(ingredients = ingredients)
 }
 case class AssemblingMachine(name: String, speed: Double, ingredientCount: Int, categories: Seq[Category]) extends AbstractItem
